Update - 0.4
Added more to the main method. All income brackets are now operational. Each income bracket will produce a message stated the income put in by the user along with their total yearly tax  utilizing "if" statements. Line 34 however contains an error that i am unable to correct nor understand the code looks correct, however the error message stating " 'else' with 'if' ", however as seen in the code this has been formatted correctly, so i am unsure as to the root of the error message.